requirements.txt
streamlit
langchain
langchain_openai
pandas
langchain_experimental
openai
tabulate
python-dotenv==1.0.1